tests/integration/requirements.txt
pytest>=8.0
requests>=2.31
Pillow>=10.0
reportlab>=4.0
